Improve expression language customization
