+++ Tests/ExpressionLanguage/ExpressionFunction/Security/HasAnyRoleTest.php
@@ -1,0 +1,28 @@
+<?php
+
+namespace Overblog\GraphQLBundle\Tests\ExpressionLanguage\ExpressionFunction\Security;
+
+use Overblog\GraphQLBundle\ExpressionLanguage\ExpressionFunction\Security\HasAnyRole;
+use Overblog\GraphQLBundle\Tests\ExpressionLanguage\TestCase;
+
+class HasAnyRoleTest extends TestCase
+{
+    protected function getFunctions()
+    {
+        return [new HasAnyRole()];
+    }
+
+    public function testHasAnyRole()
+    {
+        $this->assertExpressionCompile('hasAnyRole(["ROLE_ADMIN", "ROLE_USER"])', 'ROLE_ADMIN');
+
+        $this->assertExpressionCompile(
+            'hasAnyRole(["ROLE_ADMIN", "ROLE_USER"])',
+            $this->matchesRegularExpression('/^ROLE_(USER|ADMIN)$/'),
+            [],
+            $this->exactly(2),
+            false,
+            'assertFalse'
+        );
+    }
+}
