+++ Tests/ExpressionLanguage/ExpressionFunction/Security/IsAnonymousTest.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace Overblog\GraphQLBundle\Tests\ExpressionLanguage\ExpressionFunction\Security;
+
+use Overblog\GraphQLBundle\ExpressionLanguage\ExpressionFunction\Security\IsAnonymous;
+use Overblog\GraphQLBundle\Tests\ExpressionLanguage\TestCase;
+
+class IsAnonymousTest extends TestCase
+{
+    protected function getFunctions()
+    {
+        return [new IsAnonymous()];
+    }
+
+    public function testIsAnonymous()
+    {
+        $this->assertExpressionCompile('isAnonymous()', 'IS_AUTHENTICATED_ANONYMOUSLY');
+    }
+}
