+++ Tests/ExpressionLanguage/ExpressionFunction/Security/IsAuthenticatedTest.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace Overblog\GraphQLBundle\Tests\ExpressionLanguage\ExpressionFunction\Security;
+
+use Overblog\GraphQLBundle\ExpressionLanguage\ExpressionFunction\Security\IsAuthenticated;
+use Overblog\GraphQLBundle\Tests\ExpressionLanguage\TestCase;
+
+class IsAuthenticatedTest extends TestCase
+{
+    protected function getFunctions()
+    {
+        return [new IsAuthenticated()];
+    }
+
+    public function testIsAuthenticated()
+    {
+        $this->assertExpressionCompile('isAuthenticated()', $this->matchesRegularExpression('/^IS_AUTHENTICATED_(REMEMBERED|FULLY)$/'));
+    }
+}
